fix: move configuration directory to ./conf
