+++ demo/addissuer.js
@@ -6,7 +6,7 @@
 // issuer and purse to an individual wallet.
 
 import { E } from '@agoric/eventual-send';
-import dappConstants from '../ui.old/public/conf/defaults';
+import dappConstants from '../conf/defaults';
 
 // deploy.js runs in an ephemeral Node.js outside of swingset. The
 // spawner runs within ag-solo, so is persistent.  Once the deploy.js
